\name{calcu_cstat}
\alias{calcu_cstat}
\title{calcu cstat}
\description{function for calculating the c stat}
\usage{calcu_cstat(eset, Sj)}
\arguments{
  \item{eset}{number of patients in the test set }
  \item{Sj}{risk score of patients generated by other functions}
}

\value{returns c statistics}

\author{Yuqing Zhang}





