\name{coxmatC}
\alias{coxmatC}
\title{coxmatC}
\description{method for handling ties (alternative 'breslow')}
\usage{coxmatC(X, time, status)}
\arguments{
  \item{X}{
}
  \item{time}{
}
  \item{status}{
}
}



\author{Levi Waldron, Christoph Bernau, Yuqing Zhang}





