\name{fastCox}
\alias{fastCox}
\title{fastCox}
\description{use learningset only and sort according to time}
\usage{fastCox(X, y, learnind, criterion, ...)}
\arguments{
  \item{X}{
}
  \item{y}{
}
  \item{learnind}{
}
  \item{criterion}{
}
  \item{\dots}{
}
}



\author{Levi Waldron, Christoph Bernau, Yuqing Zhang}





