\name{masomenos}
\alias{masomenos}
\title{masomenos}
\description{Mas-o-menos algorithm, mainly used as training function}
\usage{masomenos(X, y, n.genes, p.cutoff)}
\arguments{
  \item{X}{mostly assayData of an ExpressionSet object }
  \item{y}{response, using two columns "dmfs.time" and "dmfs.cens" in the phenoData of an ExpressionSet object}
  \item{n.genes}{number of gene expression variables}
  \item{p.cutoff}{p threshold ??}
}

\value{return betas}

\author{Yuqing Zhang}





