\name{simulatorZ-package}
\alias{simulatorZ-package}
\alias{simulatorZ}
\docType{package}
\title{Simulator for Collections of Independent Genomic Data Sets}
\description{simulatorZ is a package intended primarily to simulate collections of
independent genomic data sets, as well as performing training and validation with
predicting algorithms. It supports ExpressionSets and SummarizedExperiment objects.}
\details{
\tabular{ll}{Package: \tab simulatorZ\cr
Type: \tab Package\cr
Title: \tab Simulator for Collections of Independent Genomic Data Sets\cr
Version: \tab 0.99.1\cr
Date: \tab 2014-08-03\cr
Author: \tab Yuqing Zhang, Christoph Bernau, Levi Waldron\cr
Maintainer: \tab Yuqing Zhang <zhangyuqing.pkusms@gmail.com>\cr
License: \tab Artistic-2.0\cr
Depends: \tab R (>= 3.1), Biobase, survival, CoxBoost\cr
Imports: \tab GenomicRanges, gbm, Hmisc\cr
Suggests: \tab BiocStyle, curatedOvarianData, parathyroidSE, superpc\cr
URL: \tab https://github.com/zhangyuqing/simulatorZ\cr
BugReports: \tab https://github.com/zhangyuqing/simulatorZ\cr
biocViews: \tab Survival\cr}
}
\author{Yuqing Zhang, Christoph Bernau, Levi Waldron}

\keyword{ package }


