\name{simulatorZ-package}
\alias{simulatorZ-package}
\alias{simulatorZ}
\docType{package}
\title{perform bootstrap simulation on Expression Sets}
\description{this package performs non-parametric and parametric bootstrap on Expression Sets}
\details{
\tabular{ll}{Package: \tab simulatorZ\cr
Type: \tab Package\cr
Title: \tab perform bootstrap simulation on Expression Sets\cr
Version: \tab 0.99.0\cr
Date: \tab 2014-08-03\cr
Author: \tab Levi Waldron, Christoph Bernau, Yuqing Zhang\cr
Maintainer: \tab Yuqing Zhang <zhangyuqing.pkusms@gmail.com>\cr
License: \tab Artistic-2.0\cr
Depends: \tab R (>= 3.1), Biobase, survival, CoxBoost, gbm, Hmisc\cr
Imports: \tab GenomicRanges\cr
Suggests: \tab curatedOvarianData, GenomicRanges\cr
URL: \tab https://github.com/zhangyuqing/simulatorZ\cr
BugReports: \tab https://github.com/zhangyuqing/simulatorZ\cr
biocViews: \tab Survival\cr}
}
\author{Levi Waldron, Christoph Bernau, Yuqing Zhang}

\keyword{ package }


