\name{simulatorZ-package}
\alias{simulatorZ-package}
\alias{simulatorZ}
\docType{package}
\title{simulate the result matrix for cross validation and cross study validation.}
\description{R scripts for testing whether shifts of covariates will impact prediction accuracy in cross study validation}
\details{
\tabular{ll}{Package: \tab simulatorZ\cr
Type: \tab Package\cr
Title: \tab simulate the result matrix for cross validation and cross study validation.\cr
Version: \tab 0.99.0\cr
Date: \tab 2014-08-03\cr
Author: \tab Yuqing Zhang\cr
Maintainer: \tab Yuqing Zhang <zhangyuqing.pkusms@gmail.com>\cr
License: \tab Artistic-2.0\cr}
}
\author{Yuqing Zhang}

\keyword{ package }


